+++ Sources/RESTLayer/Service/RestService.swift
@@ -18,76 +18,25 @@ public class RestService {
     ///Array of aditional HTTP header fields.
     private let headerFields: [ApiHeader]?
 
+    ///Provider of decoder and encoder.
+    private let coder: CoderProvider
+
     ///Service for managing request with REST server.
     internal let apiService: ApiService
-
-    ///Creates full url by joining *baseUrl* and *apiPath*.
-    fileprivate func requestUrl(for resourceName: String) -> URL {
-        if let apiPath = apiPath {
-            return baseUrl.appendingPathComponent(apiPath).appendingPathComponent(resourceName)
-        }
-        return baseUrl.appendingPathComponent(resourceName)
-    }
-
-    ///Merges *headerFields* with aditional headers.
-    fileprivate func apiHeaders(adding headers: [ApiHeader]?) -> [ApiHeader]? {
-        guard let headers = headers else {
-            return headerFields
-        }
-        var mergedHeders = headerFields
-        mergedHeders?.append(contentsOf: headers)
-        return mergedHeders
-    }
-
-    ///Converts RestDataResponseCimpletionHandler into ApiResponseCompletionHandler
-    fileprivate func completionHandler(for resource: RestDataResource, with restHandler: RestDataResponseCompletionHandler?) -> ApiResponseCompletionHandler? {
-        guard let restHandler = restHandler else {
-            return nil
-        }
-        return { (response: ApiResponse?, error: Error?) in
-            var resource = resource
-            var errorResponse: RestErrorResponse?
-            if let internalError = error {
-                errorResponse = RestErrorResponse(error: internalError)
-            } else if let r = response, let apiError = ApiError.error(for: r.statusCode) {
-                errorResponse = RestErrorResponse(error: apiError)
-            } else if let r = response, let parsingError = resource.update(with: r.body, aditionalInfo: RestResponseHeader.list(with: r.allHeaderFields)) {
-                errorResponse = RestErrorResponse(error: parsingError)
-            }
-            restHandler(resource, errorResponse)
-        }
-    }
-
-    ///Converts RestFileResponseCompletionHandler into ApiResponseCompletionHandler
-    fileprivate func completionHandler(for resource: RestFileResource, with restHandler: RestFileResponseCompletionHandler?) -> ApiResponseCompletionHandler? {
-        guard let restHandler = restHandler else {
-            return nil
-        }
-        return { (response: ApiResponse?, error: Error?) in
-            var resource = resource
-            var errorResponse: RestErrorResponse?
-            if let internalError = error {
-                errorResponse = RestErrorResponse(error: internalError)
-            } else if let r = response, let apiError = ApiError.error(for: r.statusCode) {
-                errorResponse = RestErrorResponse(error: apiError)
-            } else if let r = response, let updatingError = resource.update(with: RestResponseHeader.list(with: r.allHeaderFields)) {
-                errorResponse = RestErrorResponse(error: updatingError)
-            }
-            restHandler(resource, errorResponse)
-        }
-    }
 
     /**
      - Parameters:
        - baseUrl: Base URL of API server.
        - apiPath: Path of API on server.
        - headerFields: Array of HTTP header fields which will be added to all requests.
-       - fileManager: Object of class implementing *FileManagerProtocol*.
+       - coderProvider: Object providing *JSONCoder* and *JSONDecoder*.
+       - fileManager: Object of class implementing *FileManager* Protocol.
      */
-    public init(baseUrl: URL, apiPath: String? = nil, headerFields: [ApiHeader]? = nil, fileManager: FileManagerProtocol = FileCommander()) {
+    public init(baseUrl: URL, apiPath: String? = nil, headerFields: [ApiHeader]? = nil, coderProvider: CoderProvider = DefaultCoderProvider(), fileManager: FileManager = DefaultFileManager()) {
         self.baseUrl = baseUrl
         self.apiPath = apiPath
         self.headerFields = headerFields
+        self.coder = coderProvider
         self.apiService = ApiService(fileManager: fileManager)
     }
 }
@@ -96,88 +45,152 @@ public class RestService {
 public extension RestService {
 
     /**
-     Sends HTTP GET request for given resource.
+     Sends HTTP GET request.
      - Parameters:
-       - resource: RestResource object which should be filled up with response data.
+       - type: Type conforming *Decodable* protocol which should be returned in completion block.
+       - path: Path of the resource.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
     @discardableResult
-    func get(resource: RestDataResource, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestDataResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
+    func get<Response: Decodable>(type: Response.Type, from path: ResourcePath, with aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestResponseCompletionHandler<Response>? = nil) -> ApiRequest {
+        let url = requestUrl(for: path)
         let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.getData(from: url, with: headers, useProgress: useProgress, completion: handler)
+        let completion = completionHandler(for: type, coder: coder, with: completion)
+        return apiService.getData(from: url, with: headers, useProgress: useProgress, completion: completion)
     }
 
     /**
-     Sends HTTP POST request with given resource.
+     Sends HTTP POST request.
      - Parameters:
-       - resource: RestResource object which should be send and filled up with response data.
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - path: Path of the resource.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
+       - responseType: Type conforming *Decodable* protocol which should be returned in completion block.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
     @discardableResult
-    func post(resource: RestDataResource, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestDataResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
-        let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.post(data: resource.data, at: url, with: headers, useProgress: useProgress, completion: handler)
+    func post<Request: Encodable, Response: Decodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, responseType: Response.Type? = nil, completion: RestResponseCompletionHandler<Response>? = nil) -> ApiRequest {
+        let completion = completionHandler(for: Response.self, coder: coder, with: completion)
+        return post(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
     }
 
     /**
-     Sends HTTP PUT request with given resource.
+     Sends HTTP POST request.
      - Parameters:
-       - resource: RestResource object which should be send and filled up with response data.
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - path: Path of the resource.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
     @discardableResult
-    func put(resource: RestDataResource, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestDataResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
-        let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.put(data: resource.data, at: url, with: headers, useProgress: useProgress, completion: handler)
+    func post<Request: Encodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let completion = completionHandler(with: completion)
+        return post(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
     }
 
     /**
-     Sends HTTP PATCH request with given resource.
+     Sends HTTP PUT request.
      - Parameters:
-       - resource: RestResource object which should be send and filled up with response data.
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - path: Path of the resource.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
+       - responseType: Type conforming *Decodable* protocol which should be returned in completion block.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
     @discardableResult
-    func patch(resource: RestDataResource, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestDataResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
-        let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.patch(data: resource.data, at: url, with: headers, useProgress: useProgress, completion: handler)
+    func put<Request: Encodable, Response: Decodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, responseType: Response.Type? = nil, completion: RestResponseCompletionHandler<Response>? = nil) -> ApiRequest {
+        let completion = completionHandler(for: Response.self, coder: coder, with: completion)
+        return put(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
     }
 
     /**
-     Sends HTTP DELETE request for given resource.
+     Sends HTTP PUT request.
      - Parameters:
-       - resource: RestResource object which should be filled up with response data.
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - path: Path of the resource.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
     @discardableResult
-    func delete(resource: RestDataResource, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestDataResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
-        let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.delete(data: resource.data, at: url, with: headers, useProgress: useProgress, completion: handler)
+    func put<Request: Encodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let completion = completionHandler(with: completion)
+        return put(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
+    }
+
+    /**
+     Sends HTTP PATCH request.
+     - Parameters:
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - aditionalHeaders: Additional header fields which should be sent with request.
+       - useProgress: Flag indicates if Progress object should be created.
+       - responseType: Type conforming *Decodable* protocol which should be returned in completion block.
+       - completion: Closure called when request is finished.
+     - Returns: ApiRequest object which allows to follow progress and manage request.
+     */
+    @discardableResult
+    func patch<Request: Encodable, Response: Decodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, responseType: Response.Type? = nil, completion: RestResponseCompletionHandler<Response>? = nil) -> ApiRequest {
+        let completion = completionHandler(for: Response.self, coder: coder, with: completion)
+        return patch(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
+    }
+
+    /**
+     Sends HTTP PATCH request.
+     - Parameters:
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - path: Path of the resource.
+       - aditionalHeaders: Additional header fields which should be sent with request.
+       - useProgress: Flag indicates if Progress object should be created.
+       - completion: Closure called when request is finished.
+     - Returns: ApiRequest object which allows to follow progress and manage request.
+     */
+    @discardableResult
+    func patch<Request: Encodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let completion = completionHandler(with: completion)
+        return patch(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
+    }
+
+    /**
+     Sends HTTP DELETE request.
+     - Parameters:
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - path: Path of the resource.
+       - aditionalHeaders: Additional header fields which should be sent with request.
+       - useProgress: Flag indicates if Progress object should be created.
+       - responseType: Type conforming *Decodable* protocol which should be returned in completion block.
+       - completion: Closure called when request is finished.
+     - Returns: ApiRequest object which allows to follow progress and manage request.
+     */
+    @discardableResult
+    func delete<Request: Encodable, Response: Decodable>(_ value: Request? = nil, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, responseType: Response.Type? = nil, completion: RestResponseCompletionHandler<Response>? = nil) -> ApiRequest {
+        let completion = completionHandler(for: Response.self, coder: coder, with: completion)
+        return delete(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
+    }
+
+    /**
+     Sends HTTP DELETE request.
+     - Parameters:
+       - value: Value of type conforming *Encodable* protocol which data should be sent with request.
+       - path: Path of the resource.
+       - aditionalHeaders: Additional header fields which should be sent with request.
+       - useProgress: Flag indicates if Progress object should be created.
+       - completion: Closure called when request is finished.
+     - Returns: ApiRequest object which allows to follow progress and manage request.
+     */
+    @discardableResult
+    func delete<Request: Encodable>(_ value: Request? = nil, at path: ResourcePath, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = false, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let completion = completionHandler(with: completion)
+        return delete(value, at: path, aditionalHeaders: aditionalHeaders, useProgress: useProgress, completion: completion)
     }
 }
 
@@ -185,67 +198,75 @@ public extension RestService {
 public extension RestService {
 
     /**
-     Sends HTTP GET request for given resource.
+     Sends HTTP GET request for file.
      - Parameters:
-       - resource: RestResource object which should be filled up with response data.
+       - path: Path of file to download.
+       - destinationUrl: Local url, where file should be saved.
+       - inBackground: Flag indicates if downloading should be performed in background or foreground.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
-    func getFile(resource: RestFileResource, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestFileResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
+    func getFile(at path: ResourcePath, saveAt destinationUrl: URL, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let url = requestUrl(for: path)
         let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.downloadFile(from: url, to: resource.location, with: headers, inBackground: inBackground, useProgress: useProgress, completion: handler)
+        let completion = completionHandler(with: completion)
+        return apiService.downloadFile(from: url, to: destinationUrl, with: headers, inBackground: inBackground, useProgress: useProgress, completion: completion)
     }
 
     /**
-     Sends HTTP POST request with given resource.
+     Sends HTTP POST request with file.
      - Parameters:
-       - resource: RestResource object which should be send and filled up with response data.
+       - url: Local url, where file should be saved.
+       - path: Destination path of file.
+       - inBackground: Flag indicates if downloading should be performed in background or foreground.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
-    func postFile(resource: RestFileResource, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestFileResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
+    func postFile(from url: URL, at path: ResourcePath, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let destinationUrl = requestUrl(for: path)
         let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.postFile(from: resource.location, to: url, with: headers, inBackground: inBackground, useProgress: useProgress, completion: handler)
+        let completion = completionHandler(with: completion)
+        return apiService.postFile(from: url, to: destinationUrl, with: headers, inBackground: inBackground, useProgress: useProgress, completion: completion)
     }
 
     /**
-     Sends HTTP PUT request with given resource.
+     Sends HTTP PUT request with file.
      - Parameters:
-       - resource: RestResource object which should be send and filled up with response data.
+       - url: Local url, where file should be saved.
+       - path: Destination path of file.
+       - inBackground: Flag indicates if downloading should be performed in background or foreground.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
-    func putFile(resource: RestFileResource, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestFileResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
+    func putFile(from url: URL, at path: ResourcePath, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let destinationUrl = requestUrl(for: path)
         let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.putFile(from: resource.location, to: url, with: headers, inBackground: inBackground, useProgress: useProgress, completion: handler)
+        let completion = completionHandler(with: completion)
+        return apiService.putFile(from: url, to: destinationUrl, with: headers, inBackground: inBackground, useProgress: useProgress, completion: completion)
     }
 
     /**
-     Sends HTTP PATCH request with given resource.
+     Sends HTTP PATCH request with file.
      - Parameters:
-       - resource: RestResource object which should be send and filled up with response data.
+       - url: Local url, where file should be saved.
+       - path: Destination path of file.
+       - inBackground: Flag indicates if downloading should be performed in background or foreground.
        - aditionalHeaders: Additional header fields which should be sent with request.
        - useProgress: Flag indicates if Progress object should be created.
        - completion: Closure called when request is finished.
      - Returns: ApiRequest object which allows to follow progress and manage request.
      */
-    func patchFile(resource: RestFileResource, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestFileResponseCompletionHandler? = nil) -> ApiRequest {
-        let url = requestUrl(for: resource.name)
+    func patchFile(from url: URL, at path: ResourcePath, inBackground: Bool = true, aditionalHeaders: [ApiHeader]? = nil, useProgress: Bool = true, completion: RestSimpleResponseCompletionHandler? = nil) -> ApiRequest {
+        let destinationUrl = requestUrl(for: path)
         let headers = apiHeaders(adding: aditionalHeaders)
-        let handler = completionHandler(for: resource, with: completion)
-        return apiService.patchFile(from: resource.location, to: url, with: headers, inBackground: inBackground, useProgress: useProgress, completion: handler)
+        let completion = completionHandler(with: completion)
+        return apiService.patchFile(from: url, to: destinationUrl, with: headers, inBackground: inBackground, useProgress: useProgress, completion: completion)
     }
 }
 
@@ -255,5 +276,114 @@ public extension RestService {
     ///Cancels all currently running requests.
     func cancelAllRequests() {
         apiService.cancelAllRequests()
+    }
+}
+
+//MARK: - Parameter factories
+private extension RestService {
+
+    ///Creates full url by joining `baseUrl`, `apiPath` and given `path`.
+    func requestUrl(for path: ResourcePath) -> URL {
+        if let apiPath = apiPath, !apiPath.isEmpty {
+            return baseUrl.appendingPathComponent(apiPath).appendingPathComponent(path.rawValue)
+        }
+        return baseUrl.appendingPathComponent(path.rawValue)
+    }
+
+    ///Merges `headerFields` with giver aditional `headers`.
+    func apiHeaders(adding headers: [ApiHeader]?) -> [ApiHeader]? {
+        guard let headerFields = headerFields else {
+            return headers
+        }
+        guard let headers = headers else {
+            return headerFields
+        }
+        var mergedHeders = headerFields
+        mergedHeders.append(contentsOf: headers)
+        return mergedHeders
+    }
+
+    ///Encodes encodable `value` into *Data* object.
+    func requestData<T: Encodable>(for value: T?) -> Data? {
+        guard let value = value else {
+            return nil
+        }
+        return try? coder.encode(value)
+    }
+
+    ///Converts *RestResponseCompletionHandler* into *ApiResponseCompletionHandler*.
+    func completionHandler<Response: Decodable>(for type: Response.Type?, coder: CoderProvider, with completion: RestResponseCompletionHandler<Response>?) -> ApiResponseCompletionHandler? {
+        guard let completion = completion, let type = type else {
+            return nil
+        }
+        return { (response, error) in
+            guard let response = response, error == nil else {
+                completion(nil, error)
+                return
+            }
+            guard let body = response.body else {
+                completion(nil, ApiError.error(for: response.statusCode))
+                return
+            }
+            do {
+                let decodedData = try coder.decode(type, from: body)
+                completion(decodedData, nil)
+            } catch {
+                completion(nil, error)
+            }
+        }
+    }
+
+    ///Converts *SimpleRestResponseCompletionHandler* into *ApiResponseCompletionHandler*.
+    func completionHandler(with completion: RestSimpleResponseCompletionHandler?) -> ApiResponseCompletionHandler? {
+        guard let completion = completion else {
+            return nil
+        }
+        return { (response, error) in
+            guard let response = response, error == nil else {
+                completion(false, error)
+                return
+            }
+            guard response.statusCode.isSuccess else {
+                completion(false, ApiError.error(for: response.statusCode))
+                return
+            }
+            completion(true, nil)
+        }
+    }
+}
+
+private extension RestService {
+
+    ///Posts given `value` using `apiService`.
+    func post<Request: Encodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]?, useProgress: Bool, completion: ApiResponseCompletionHandler?) -> ApiRequest {
+        let url = requestUrl(for: path)
+        let data = requestData(for: value)
+        let headers = apiHeaders(adding: aditionalHeaders)
+        return apiService.post(data: data, at: url, with: headers, useProgress: useProgress, completion: completion)
+    }
+
+    ///Puts given `value` using `apiService`.
+    func put<Request: Encodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]?, useProgress: Bool, completion: ApiResponseCompletionHandler?) -> ApiRequest {
+        let url = requestUrl(for: path)
+        let data = requestData(for: value)
+        let headers = apiHeaders(adding: aditionalHeaders)
+        return apiService.put(data: data, at: url, with: headers, useProgress: useProgress, completion: completion)
+    }
+
+    ///Patches given `value` using `apiService`.
+    func patch<Request: Encodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]?, useProgress: Bool, completion: ApiResponseCompletionHandler?) -> ApiRequest {
+        let url = requestUrl(for: path)
+        let data = requestData(for: value)
+        let headers = apiHeaders(adding: aditionalHeaders)
+        return apiService.patch(data: data, at: url, with: headers, useProgress: useProgress, completion: completion)
+    }
+
+    ///Deletes given `value` using `apiService`.
+    func delete<Request: Encodable>(_ value: Request?, at path: ResourcePath, aditionalHeaders: [ApiHeader]?, useProgress: Bool, completion: ApiResponseCompletionHandler?) -> ApiRequest {
+        let url = requestUrl(for: path)
+        let data = requestData(for: value)
+        let headers = apiHeaders(adding: aditionalHeaders)
+        return apiService.delete(data: data, at: url, with: headers, useProgress: useProgress, completion: completion)
     }
 }